Make SPOC gradebook an API implementation
rather than just hacked in to the instructor_dashboard base file.
Also move the tests from the legacy implementation to the new dash API implementation.
