+++ lms/djangoapps/instructor/views/api.py
@@ -31,7 +31,7 @@ from django_comment_common.models import (
     FORUM_ROLE_MODERATOR,
     FORUM_ROLE_COMMUNITY_TA,
 )
-
+from edxmako.shortcuts import render_to_response
 from courseware.models import StudentModule
 from student.models import CourseEnrollment, unique_id_for_user, anonymous_id_for_user
 import instructor_task.api
@@ -46,6 +46,7 @@ from instructor.enrollment import (
     unenroll_email
 )
 from instructor.access import list_with_level, allow_access, revoke_access, update_forum_role
+from instructor.offline_gradecalc import student_grades
 import analytics.basic
 import analytics.distributions
 import analytics.csvs
@@ -1306,3 +1307,43 @@ def _split_input_list(str_list):
     new_list = [s for s in new_list if s != '']
 
     return new_list
+
+
+#---- Gradebook (shown to small courses only) ----
+@cache_control(no_cache=True, no_store=True, must_revalidate=True)
+@require_level('staff')
+def spoc_gradebook(request, course_id):
+    """
+    Show the gradebook for this course:
+    - Only shown for courses with enrollment < settings.FEATURES.get("MAX_ENROLLMENT_INSTR_BUTTONS")
+    - Only displayed to course staff
+    """
+    course_key = SlashSeparatedCourseKey.from_deprecated_string(course_id)
+    course = get_course_with_access(request.user, 'staff', course_key, depth=None)
+
+    enrolled_students = User.objects.filter(
+        courseenrollment__course_id=course_key,
+        courseenrollment__is_active=1
+    ).order_by('username').select_related("profile")
+
+    # possible extension: implement pagination to show to large courses
+
+    student_info = [
+        {
+            'username': student.username,
+            'id': student.id,
+            'email': student.email,
+            'grade_summary': student_grades(student, request, course),
+            'realname': student.profile.name,
+        }
+        for student in enrolled_students
+    ]
+
+    return render_to_response('courseware/gradebook.html', {
+        'students': student_info,
+        'course': course,
+        'course_id': course_key,
+        # Checked above
+        'staff_access': True,
+        'ordered_grades': sorted(course.grade_cutoffs.items(), key=lambda i: i[1], reverse=True),
+    })
